requirements.txt
Django==1.10.5
djangorestframework==3.5.3
django-filter==1.0.1
